move fp4 utils to export folder
